server/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
pydantic[email]
passlib[bcrypt]
PyJWT
httpx
